fix: Adjust formula to generating damage attack and balance some pokemon stats

+++ Assets/Scripts/Pokemon/Data/PokemonData.cs
@@ -7,7 +7,7 @@ public class PokemonData : ScriptableObject
 
     public int MaxHealth;
     public int Attack;
-    public float Defense;
+    public int Defense;
 
     public Sprite sprite;
